# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all business profiles
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{businessId}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business profile
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business profile
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business profile
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Products ─────────────────────────────────────────────────────────────────
...
  /blueprints/{blueprintId}:
    get:
      operationId: getBlueprint
      tags: [blueprints]
      summary: Get a blueprint
      parameters:
        - name: blueprintId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blueprint with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blueprint"
        "404":
          description: Not found

  # ─── Campaigns ────────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{campaignId}/approve:
    post:
      operationId: approveCampaign
      tags: [campaigns, approvals]
      summary: Approve a campaign (human gate)
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecision"
      responses:
        "200":
          description: Campaign approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/adsets:
    get:
      operationId: listAdSets
      tags: [adsets]
      summary: List ad sets for a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of ad sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdSet"
    post:
      operationId: createAdSet
      tags: [adsets]
      summary: Create an ad set for a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdSetInput"
      responses:
        "201":
          description: Ad set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
  /adsets/{adsetId}:
    get:
      operationId: getAdSet
      tags: [adsets]
      summary: Get an ad set
      parameters:
        - name: adsetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
        "404":
          description: Not found
    patch:
      operationId: updateAdSet
      tags: [adsets]
      summary: Update an ad set
      parameters:
        - name: adsetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdSetUpdate"
      responses:
        "200":
          description: Ad set updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
    delete:
      operationId: deleteAdSet
      tags: [adsets]
      summary: Delete an ad set
      parameters:
        - name: adsetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Business:
      type: object
      required: [id, businessName, industry, createdAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        industry:
          type: string
        productCategory:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        targetMarket:
          type: ["string", "null"]
        valueProposition:
          type: ["string", "null"]
        brandVoice:
          type: ["string", "null"]
        complianceNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    BusinessInput:
      type: object
      required: [businessName, industry]
      properties:
        businessName:
          type: string
          minLength: 1
        industry:
          type: string
          minLength: 1
        productCategory:
          type: string
        location:
          type: string
        website:
          type: string
        whatsapp:
          type: string
        targetMarket:
          type: string
        valueProposition:
          type: string
        brandVoice:
          type: string
        complianceNotes:
          type: string
    BusinessUpdate:
      type: object
      properties:
        businessName:
          type: string
        industry:
          type: string
        productCategory:
          type: string
        location:
          type: string
        website:
          type: string
        whatsapp:
          type: string
        targetMarket:
          type: string
        valueProposition:
          type: string
        brandVoice:
          type: string
        complianceNotes:
          type: string

    # ─── Product ─────────────────────────────────────────────────────────────────
...
    Blueprint:
      type: object
      required: [id, pipelineRunId, title, createdAt]
      properties:
        id:
          type: integer
        pipelineRunId:
          type: integer
        title:
          type: string
        businessContext:
          type: ["string", "null"]
          description: Business diagnosis section (JSON string)
        campaignStrategy:
          type: ["string", "null"]
          description: Campaign strategy recommendation (JSON string)
        audiencePlan:
          type: ["string", "null"]
          description: Audience intelligence blueprint (JSON string)
        offerStrategy:
          type: ["string", "null"]
          description: Offer engineering blueprint (JSON string)
        creativeBlueprint:
          type: ["string", "null"]
          description: Creative copy and angle blueprint (JSON string)
        budgetPlan:
          type: ["string", "null"]
          description: Budget and testing plan (JSON string)
        policyReview:
          type: ["string", "null"]
          description: Compliance and policy risk assessment (JSON string)
        conversionReadinessScore:
          type: ["integer", "null"]
          description: 0-100 score
        policyRiskScore:
          type: ["integer", "null"]
          description: 0-100 score
        creativeStrengthScore:
          type: ["integer", "null"]
          description: 0-100 score
        funnelFitScore:
          type: ["integer", "null"]
          description: 0-100 score
        approvalStatus:
          type: string
          description: "draft | needs_revision | approved | published"
        createdAt:
          type: string

    # ─── Campaign ────────────────────────────────────────────────────────────────
    Campaign:
      type: object
      required: [id, businessId, campaignName, objective, status, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        metaCampaignId:
          type: ["string", "null"]
        blueprintId:
          type: ["integer", "null"]
        campaignName:
          type: string
        objective:
          type: string
          description: "AWARENESS | TRAFFIC | ENGAGEMENT | LEADS | APP_PROMOTION | SALES"
        status:
          type: string
          description: "draft | review | approved | active | learning | optimizing | scaling | paused | completed"
        budgetType:
          type: ["string", "null"]
          description: "daily | lifetime"
        dailyBudget:
          type: ["number", "null"]
        lifetimeBudget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        approvalStatus:
          type: string
          description: "pending | approved | rejected"
        specialAdCategory:
          type: ["boolean", "null"]
        campaignBudgetOptimization:
          type: ["boolean", "null"]
        createdByAgent:
          type: ["boolean", "null"]
        approvedByUser:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CampaignInput:
      type: object
      required: [businessId, campaignName, objective]
      properties:
        businessId:
          type: integer
        blueprintId:
          type: integer
        campaignName:
          type: string
          minLength: 1
        objective:
          type: string
        budgetType:
          type: string
        dailyBudget:
          type: number
        lifetimeBudget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        specialAdCategory:
          type: boolean
        campaignBudgetOptimization:
          type: boolean
    CampaignUpdate:
      type: object
      properties:
        campaignName:
          type: string
        objective:
          type: string
        status:
          type: string
        budgetType:
          type: string
        dailyBudget:
          type: number
        lifetimeBudget:
          type: number
        startDate:
          type: string
        endDate:
          type: string

    # ─── Ad Set ──────────────────────────────────────────────────────────────────
    AdSet:
      type: object
      required: [id, campaignId, adsetName, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        metaAdsetId:
          type: ["string", "null"]
        adsetName:
          type: string
        audienceName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        ageMin:
          type: ["integer", "null"]
        ageMax:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        interests:
          type: ["string", "null"]
          description: JSON stringified array of interests
        customAudience:
          type: ["string", "null"]
        lookalikAudience:
          type: ["string", "null"]
        placement:
          type: ["string", "null"]
          description: "advantage_plus | manual"
        optimizationEvent:
          type: ["string", "null"]
          description: "LEAD | PURCHASE | ADD_TO_CART | CONTACT"
        attributionSetting:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        conversionLocation:
          type: ["string", "null"]
          description: "WEBSITE | APP | MESSAGING | CALLS | LEADS"
        createdAt:
          type: string
    AdSetInput:
      type: object
      required: [adsetName]
      properties:
        adsetName:
          type: string
          minLength: 1
        audienceName:
          type: string
        location:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        gender:
          type: string
        language:
          type: string
        interests:
          type: string
        placement:
          type: string
        optimizationEvent:
          type: string
        attributionSetting:
          type: string
        budget:
          type: number
        conversionLocation:
          type: string
    AdSetUpdate:
      type: object
      properties:
        adsetName:
          type: string
        audienceName:
          type: string
        location:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        gender:
          type: string
        language:
          type: string
        interests:
          type: string
        placement:
          type: string
        optimizationEvent:
          type: string
        attributionSetting:
          type: string
        budget:
          type: number
        conversionLocation:
          type: string

    # ─── Creative ────────────────────────────────────────────────────────────────
...
    ApprovalDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          description: "approved | rejected"
        notes:
          type: string

    # ─── Agent Memory ─────────────────────────────────────────────────────────────
...