# lib/api-zod/src/generated/types/adSet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface AdSet {
  id: number;
  campaignId: number;
  /** @nullable */
  metaAdsetId?: string | null;
  adsetName: string;
  /** @nullable */
  audienceName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  language?: string | null;
  /**
     * JSON stringified array of interests
     * @nullable
     */
  interests?: string | null;
  /** @nullable */
  customAudience?: string | null;
  /** @nullable */
  lookalikAudience?: string | null;
  /**
     * advantage_plus | manual
     * @nullable
     */
  placement?: string | null;
  /**
     * LEAD | PURCHASE | ADD_TO_CART | CONTACT
     * @nullable
     */
  optimizationEvent?: string | null;
  /** @nullable */
  attributionSetting?: string | null;
  /** @nullable */
  budget?: number | null;
  /**
     * WEBSITE | APP | MESSAGING | CALLS | LEADS
     * @nullable
     */
  conversionLocation?: string | null;
  createdAt: string;
}