# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AdClaw AI - Agentic Meta Ads Blueprint & Conversion Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  /** @minLength 1 */
  businessName: string;
  /** @minLength 1 */
  industry: string;
  productCategory?: string;
  location?: string;
  website?: string;
  whatsapp?: string;
  targetMarket?: string;
  valueProposition?: string;
  brandVoice?: string;
  complianceNotes?: string;
}